[expo-notifications] Rename InstallationIdProvider to ServerRegistrationModule
