+++ packages/expo-notifications/android/src/main/java/expo/modules/notifications/serverregistration/InstallationId.java
@@ -1,4 +1,4 @@
-package expo.modules.notifications.installationid;
+package expo.modules.notifications.serverregistration;
 
 import android.content.Context;
 import android.content.SharedPreferences;
