--- packages/expo-notifications/src/InstallationIdProvider.native.ts
@@ -1,5 +1,0 @@
-import { NativeModulesProxy } from '@unimodules/core';
-
-import { InstallationIdProvider } from './InstallationIdProvider.types';
-
-export default (NativeModulesProxy.NotificationsInstallationIdProvider as any) as InstallationIdProvider;
