--- packages/expo-notifications/src/InstallationIdProvider.ts
@@ -1,6 +1,0 @@
-import { InstallationIdProvider } from './InstallationIdProvider.types';
-
-export default {
-  addListener: () => {},
-  removeListeners: () => {},
-} as InstallationIdProvider;
