+++ packages/expo-notifications/src/ServerRegistrationModule.native.ts
@@ -1,0 +1,5 @@
+import { NativeModulesProxy } from '@unimodules/core';
+
+import { ServerRegistrationModule } from './ServerRegistrationModule.types';
+
+export default (NativeModulesProxy.NotificationsServerRegistrationModule as any) as ServerRegistrationModule;
