+++ packages/expo-notifications/src/ServerRegistrationModule.ts
@@ -1,0 +1,6 @@
+import { ServerRegistrationModule } from './ServerRegistrationModule.types';
+
+export default {
+  addListener: () => {},
+  removeListeners: () => {},
+} as ServerRegistrationModule;
